+++ support-frontend/assets/pages/[countryGroupId]/components/SoftOptInCheckoutConsent.tsx
